Consolidate into config struct
