+++ cli/cmd/proxy/read/config_test.go
@@ -1,63 +1,52 @@
 package read
 
 import (
+	"context"
 	"embed"
+	"net/http"
+	"net/http/httptest"
+	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/require"
 )
 
 //go:embed test_config_dump.json
-var config_fs embed.FS
+var fs embed.FS
 
-func TestParseConfig(t *testing.T) {
-	testConfig, err := config_fs.ReadFile("test_config_dump.json")
+func TestFetchConfig(t *testing.T) {
+	configResponse, err := fs.ReadFile("test_config_dump.json")
 	require.NoError(t, err)
 
-	expected := Config{
-		Clusters: []Cluster{
-			{
-				Name:                     "local_agent",
-				FullyQualifiedDomainName: "local_agent",
-				Endpoints:                []string{"192.168.79.187:8502"},
-				Type:                     "STATIC",
-				LastUpdated:              "2022-05-13T04:22:39.553Z",
-			},
-			{
-				Name:                     "client",
-				FullyQualifiedDomainName: "client.default.dc1.internal.bc3815c2-1a0f-f3ff-a2e9-20d791f08d00.consul",
-				Type:                     "EDS",
-				LastUpdated:              "2022-06-09T00:39:12.948Z",
-			},
-			{
-				Name:                     "frontend",
-				FullyQualifiedDomainName: "frontend.default.dc1.internal.bc3815c2-1a0f-f3ff-a2e9-20d791f08d00.consul",
-				Type:                     "EDS",
-				LastUpdated:              "2022-06-09T00:39:12.855Z",
-			},
-			{
-				Name:                     "local_app",
-				FullyQualifiedDomainName: "local_app",
-				Endpoints:                []string{"127.0.0.1:8080"},
-				Type:                     "STATIC",
-				LastUpdated:              "2022-05-13T04:22:39.655Z",
-			},
-			{
-				Name:                     "original-destination",
-				FullyQualifiedDomainName: "original-destination",
-				Type:                     "ORIGINAL_DST",
-				LastUpdated:              "2022-05-13T04:22:39.743Z",
-			},
-			{
-				Name:                     "server",
-				FullyQualifiedDomainName: "server.default.dc1.internal.bc3815c2-1a0f-f3ff-a2e9-20d791f08d00.consul",
-				Type:                     "EDS",
-				LastUpdated:              "2022-06-09T00:39:12.754Z",
-			},
+	mockServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.Write(configResponse)
+	}))
+	defer mockServer.Close()
+
+	mpf := &mockPortForwarder{
+		openBehavior: func(ctx context.Context) (string, error) {
+			return strings.Replace(mockServer.URL, "http://", "", 1), nil
 		},
 	}
 
-	actual, err := ParseConfig(testConfig)
+	configDump, err := FetchConfig(context.Background(), mpf)
+
 	require.NoError(t, err)
-	require.Equal(t, expected, actual)
+	require.NotNil(t, configDump)
 }
+
+func TestParseConfig(t *testing.T) {
+	testConfig, err := fs.ReadFile("test_config_dump.json")
+	require.NoError(t, err)
+
+	envoyConfig, err := NewEnvoyConfig(testConfig)
+	require.NoError(t, err)
+	require.NotNil(t, envoyConfig)
+}
+
+type mockPortForwarder struct {
+	openBehavior func(context.Context) (string, error)
+}
+
+func (m *mockPortForwarder) Open(ctx context.Context) (string, error) { return m.openBehavior(ctx) }
+func (m *mockPortForwarder) Close()                                   {}
