--- cli/cmd/proxy/read/print_test.go
@@ -1,1 +1,0 @@
-package read
